+++ server/controller.go
@@ -16,45 +16,40 @@
 package main
 
 import (
-	"bytes"
 	"encoding/base64"
 	"errors"
 	"fmt"
 	"log"
 	"os"
-	"os/exec"
 	"os/signal"
-	"path"
 	"strconv"
-	"strings"
 	"sync"
 	"time"
 )
 
 type Controller struct {
-	Admin        *Admin
-	Api          *Api
-	Calls        *Calls
-	Config       *Config
-	Database     *Database
-	Accesses     *Accesses
-	Apikeys      *Apikeys
-	Dirwatches   *Dirwatches
-	Downstreams  *Downstreams
-	Groups       *Groups
-	Logs         *Logs
-	Options      *Options
-	Scheduler    *Scheduler
-	Systems      *Systems
-	Tags         *Tags
-	Clients      *Clients
-	Register     chan *Client
-	Unregister   chan *Client
-	Ingest       chan *Call
-	ffmpeg       bool
-	ffmpegWarned bool
-	ingestMutex  sync.Mutex
-	running      bool
+	Admin       *Admin
+	Api         *Api
+	Calls       *Calls
+	Config      *Config
+	Database    *Database
+	Accesses    *Accesses
+	Apikeys     *Apikeys
+	Dirwatches  *Dirwatches
+	Downstreams *Downstreams
+	FFMpeg      *FFMpeg
+	Groups      *Groups
+	Logs        *Logs
+	Options     *Options
+	Scheduler   *Scheduler
+	Systems     *Systems
+	Tags        *Tags
+	Clients     *Clients
+	Register    chan *Client
+	Unregister  chan *Client
+	Ingest      chan *Call
+	ingestMutex sync.Mutex
+	running     bool
 }
 
 func NewController(config *Config) *Controller {
@@ -65,6 +60,7 @@ func NewController(config *Config) *Controller {
 		Calls:       NewCalls(),
 		Dirwatches:  NewDirwatches(),
 		Downstreams: NewDownstreams(),
+		FFMpeg:      NewFFMpeg(),
 		Groups:      NewGroups(),
 		Logs:        NewLogs(),
 		Options:     NewOptions(),
@@ -82,61 +78,10 @@ func NewController(config *Config) *Controller {
 	controller.Database = NewDatabase(config)
 	controller.Scheduler = NewScheduler(controller)
 
+	controller.Logs.setDaemon(config.daemon)
+	controller.Logs.setDatabase(controller.Database)
+
 	return controller
-}
-
-func (controller *Controller) ConvertAudio(call *Call) {
-	var (
-		args = []string{"-i", "-"}
-		err  error
-	)
-
-	if !controller.ffmpeg {
-		if !controller.ffmpegWarned {
-			controller.ffmpegWarned = true
-
-			controller.Logs.LogEvent(controller.Database, LogLevelWarn, "ffmpeg is not available, no audio conversion will be performed.")
-		}
-		return
-	}
-
-	if system, ok := controller.Systems.GetSystem(call.System); ok {
-		if talkgroup, ok := system.Talkgroups.GetTalkgroup(call.Talkgroup); ok {
-			if tag, ok := controller.Tags.GetTag(talkgroup.TagId); ok {
-				args = append(args,
-					"-metadata", fmt.Sprintf("album=%v", talkgroup.Label),
-					"-metadata", fmt.Sprintf("artist=%v", system.Label),
-					"-metadata", fmt.Sprintf("date=%v", call.DateTime),
-					"-metadata", fmt.Sprintf("genre=%v", tag),
-					"-metadata", fmt.Sprintf("title=%v", talkgroup.Name),
-				)
-			}
-		}
-	}
-
-	args = append(args, "-af", "apad=whole_dur=3s,loudnorm=I=-16:TP=-1.5:LRA=11", "-c:a", "aac", "-b:a", "32k", "-movflags", "frag_keyframe+empty_moov", "-f", "ipod", "-")
-
-	cmd := exec.Command("ffmpeg", args...)
-	cmd.Stdin = bytes.NewReader(call.Audio)
-
-	stdout := bytes.NewBuffer([]byte(nil))
-	cmd.Stdout = stdout
-
-	stderr := bytes.NewBuffer([]byte(nil))
-	cmd.Stderr = stderr
-
-	if err = cmd.Run(); err == nil {
-		call.Audio = stdout.Bytes()
-		call.AudioType = "audio/mp4"
-
-		switch v := call.AudioName.(type) {
-		case string:
-			call.AudioName = fmt.Sprintf("%v.m4a", strings.TrimSuffix(v, path.Ext((v))))
-		}
-
-	} else {
-		fmt.Println(stderr.String())
-	}
 }
 
 func (controller *Controller) EmitCall(call *Call) {
@@ -169,15 +114,11 @@ func (controller *Controller) IngestCall(call *Call) {
 	defer controller.IngestUnlock()
 
 	logCall := func(call *Call, level string, message string) {
-		controller.Logs.LogEvent(
-			controller.Database,
-			level,
-			fmt.Sprintf("newcall: system=%v talkgroup=%v file=%v %v", call.System, call.Talkgroup, call.AudioName, message),
-		)
+		controller.Logs.LogEvent(level, fmt.Sprintf("newcall: system=%v talkgroup=%v file=%v %v", call.System, call.Talkgroup, call.AudioName, message))
 	}
 
 	logError := func(err error) {
-		controller.Logs.LogEvent(controller.Database, LogLevelError, fmt.Sprintf("controller.ingestcall: %v", err.Error()))
+		controller.Logs.LogEvent(LogLevelError, fmt.Sprintf("controller.ingestcall: %v", err.Error()))
 	}
 
 	if system, ok = controller.Systems.GetSystem(call.System); ok {
@@ -346,7 +287,9 @@ func (controller *Controller) IngestCall(call *Call) {
 	}
 
 	if !controller.Options.DisableAudioConversion {
-		controller.ConvertAudio(call)
+		if err := controller.FFMpeg.Convert(call, controller.Systems, controller.Tags); err != nil {
+			controller.Logs.LogEvent(LogLevelWarn, err.Error())
+		}
 	}
 
 	if id, err = controller.Calls.WriteCall(call, controller.Database); err == nil {
@@ -385,11 +328,7 @@ func (controller *Controller) IngestUnlock() {
 }
 
 func (controller *Controller) LogClientsCount() {
-	controller.Logs.LogEvent(
-		controller.Database,
-		LogLevelInfo,
-		fmt.Sprintf("listeners count is %v", controller.Clients.Count()),
-	)
+	controller.Logs.LogEvent(LogLevelInfo, fmt.Sprintf("listeners count is %v", controller.Clients.Count()))
 }
 
 func (controller *Controller) ProcessMessage(client *Client, message *Message) error {
@@ -494,31 +433,19 @@ func (controller *Controller) ProcessMessageCommandPin(client *Client, message *
 			if access, ok := controller.Accesses.GetAccess(code); ok {
 				client.Access = access
 			} else {
-				controller.Logs.LogEvent(
-					controller.Database,
-					LogLevelWarn,
-					fmt.Sprintf("invalid access code=\"%s\" address=\"%s\"", code, client.GetRemoteAddr()),
-				)
+				controller.Logs.LogEvent(LogLevelWarn, fmt.Sprintf("invalid access code=\"%s\" address=\"%s\"", code, client.GetRemoteAddr()))
 				client.Send <- &Message{Command: MessageCommandPin}
 				return nil
 			}
 
 			if client.AuthCount == maxAuthCount {
-				controller.Logs.LogEvent(
-					controller.Database,
-					LogLevelWarn,
-					fmt.Sprintf("access ident=\"%s\" locked", client.Access.Ident),
-				)
+				controller.Logs.LogEvent(LogLevelWarn, fmt.Sprintf("access ident=\"%s\" locked", client.Access.Ident))
 				client.Send <- &Message{Command: MessageCommandPin}
 				return nil
 			}
 
 			if client.Access.HasExpired() {
-				controller.Logs.LogEvent(
-					controller.Database,
-					LogLevelWarn,
-					fmt.Sprintf("access ident=\"%s\" expired", client.Access.Ident),
-				)
+				controller.Logs.LogEvent(LogLevelWarn, fmt.Sprintf("access ident=\"%s\" expired", client.Access.Ident))
 				client.Send <- &Message{Command: MessageCommandExpired}
 				return nil
 			}
@@ -526,11 +453,7 @@ func (controller *Controller) ProcessMessageCommandPin(client *Client, message *
 			switch v := client.Access.Limit.(type) {
 			case uint:
 				if controller.Clients.AccessCount(client) > int(v) {
-					controller.Logs.LogEvent(
-						controller.Database,
-						LogLevelWarn,
-						fmt.Sprintf("access ident=\"%s\" too many concurrent connections, limit is %d", client.Access.Ident, client.Access.Limit),
-					)
+					controller.Logs.LogEvent(LogLevelWarn, fmt.Sprintf("access ident=\"%s\" too many concurrent connections, limit is %d", client.Access.Ident, client.Access.Limit))
 					client.Send <- &Message{Command: MessageCommandMax}
 					return nil
 				}
@@ -554,7 +477,7 @@ func (controller *Controller) Start() error {
 		controller.running = true
 	}
 
-	controller.Logs.LogEvent(controller.Database, LogLevelWarn, "server started")
+	controller.Logs.LogEvent(LogLevelWarn, "server started")
 
 	if len(controller.Config.BaseDir) > 0 {
 		log.Printf("base folder is %s\n", controller.Config.BaseDir)
@@ -590,14 +513,6 @@ func (controller *Controller) Start() error {
 	}
 	if err = controller.Scheduler.Start(); err != nil {
 		return err
-	}
-
-	cmd := exec.Command("ffmpeg", "-version")
-	if err := cmd.Run(); err == nil {
-		controller.ffmpeg = true
-
-	} else {
-		controller.ffmpeg = false
 	}
 
 	go func() {